requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
slowapi==0.1.9
sqlalchemy==2.0.41
requests==2.32.3
python-dotenv==1.1.0
langchain==0.3.26
langchain-community==0.3.24
langchain-core==0.3.70
langchain-openai==0.3.28
openai==1.97.0
pydantic==2.11.4
pydantic-core==2.33.2
typing-extensions==4.13.2

